# swagger.yaml
openapi: 3.0.0
info:
  title: BusitPlus API
  version: 1.0.0
  description: >
    BusitPlus API documentation for the BusitPlus system.

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /api/test/encrypt:
    post:
      summary: Encrypt Password
      description: Encrypts a plain text password using Bcrypt and returns the hashed value.
      tags:
        - Test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to be encrypted.
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to be encrypted.
      responses:
        '200':
          description: Password encrypted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "$2b$14$p.Fv7a5PbQVrwnSgQGFfHe96gpZKB0xRPIY.PN4FvWyfvhHC08gR2"
                    description: The Bcrypt hashed password.
                  status:
                    type: boolean
                    example: true
                    description: Indicates success.
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password is required."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "This API is not allowed in production."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.

  /api/test/decrypt:
    post:
      summary: Decrypt Password
      description: Decrypts a plain text password against a Bcrypt hash.
      tags:
        - Test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - hash
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to decrypt.
                hash:
                  type: string
                  example: "$2b$14$p.Fv7a5PbQVrwnSgQGFfHe96gpZKB0xRPIY.PN4FvWyfvhHC08gR2"
                  description: The Bcrypt hashed password to compare against.
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - password
                - hash
              properties:
                password:
                  type: string
                  example: mypassword123
                  description: The plain text password to decrypt.
                hash:
                  type: string
                  example: "$2b$14$p.Fv7a5PbQVrwnSgQGFfHe96gpZKB0xRPIY.PN4FvWyfvhHC08gR2"
                  description: The Bcrypt hashed password to compare against.
      responses:
        '200':
          description: Password decrypted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The password is correct."
                    description: Verification message.
                  status:
                    type: boolean
                    example: true
                    description: Indicates success.
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password and hash are required."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "This API is not allowed in production."
                      description: Error message.
                    status:
                      type: boolean
                      example: false
                      description: Indicates failure.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.

  /api/verifyToken:
    post:
      summary: Verify Token
      description: Verifies a JWT token from the `x-access-token` header and returns the decoded payload.
      tags:
        - Verify
      parameters:
        - name: x-access-token
          in: header
          required: true
          schema:
            type: string
            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          description: The JWT token to verify.
      responses:
        '200':
          description: Token verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token is valid."
                    description: Verification message.
                  status:
                    type: boolean
                    example: true
                    description: Indicates success.
        '201':
          description: Token verification failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token is invalid."
                    description: Verification message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '401':
          description: token not provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token is required for authentication." 
                    description: token not provided message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.
        '402': 
          description: Token invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid Token."
                    description: Error message.
                  status:
                    type: boolean
                    example: false
                    description: Indicates failure.

  /api/login/application:
    post:
      summary: Application login 
      description: Application Login a user with email and password, and return a JWT token.
      tags:
        - Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Users_Email
                - Users_Password
              properties:
                Users_Email:
                  type: string
                  example: "student.admin@rmutto.ac.th"
                Users_Password:
                  type: string
                  example: "Cs12345678"
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - Users_Email
                - Users_Password
              properties:
                Users_Email:
                  type: string
                  example: "student.admin@rmutto.ac.th"
                Users_Password:
                  type: string
                  example: "Cs12345678"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    description: JWT token for authentication
                  message:
                    type: string
                    example: "The login was successful."
                  status:
                    type: boolean
                    example: true
        '400':
          description: Invalid input parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Please fill in the correct parameters as required."
                  status:
                    type: boolean
                    example: false
        '401':
          description: Incorrect email or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The password is incorrect."
                  status:
                    type: boolean
                    example: false
        '404':
          description: Invalid user type
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User type details not found."
                  status:
                    type: boolean
                    example: false
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An error occurred on the server."
                  status:
                    type: boolean
                    example: false